config and auth to not be pasrt of testing
